requirements.txt
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Jinja2==3.0.1
MarkupSafe==2.0.1
SQLAlchemy==1.4.19
WTForms==2.3.3
Werkzeug==2.0.1
blinker==1.4
click==8.0.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
pip==21.1.3
setuptools==57.0.0
visitor==0.1.3
gunicorn==20.1.0
psycopg2-binary==2.9.1